Allow metadata for :_content, including new :include metadata.
By default, fields are still filtered for inclusion in :_content based
on whether they are stored. But this can be overriden by providing
metadata on :_content with an :include field, which is a set of fields
to include. This could be used to, for example, exclude fields that
are not being indexed.

Providing metadata on :content can also be used to tweak how it is
indexed and stored. For example, you might want to avoid storing :_content
if you do not need highlighting and want to save space.