Accept writer/reader for efficienty. Make index-* fns public.
Anywhere an index was taken, (`add`, `delete`, `search`,
`search-and-delete`), also optionally take a writer or reader as
appropriate in place of the index for situations where writes/reads
need to be done in bulk and efficiently.

Make `index-writer` and `index-reader` public fns to allow easy
creation of writers and readers.